requirements.txt
allure-pytest         2.13.5
allure-python-commons 2.13.5
attrs                 23.2.0
certifi               2024.7.4
cffi                  1.16.0
charset-normalizer    3.3.2
colorama              0.4.6
constants             0.6.0
h11                   0.14.0
idna                  3.7
iniconfig             2.0.0
outcome               1.3.0.post0
packaging             24.1
pip                   23.2.1
pluggy                1.5.0
pycparser             2.22
PySocks               1.7.1
pytest                8.2.2
python-dateutil       2.9.0.post0
requests              2.32.3
selenium              4.22.0
six                   1.16.0
sniffio               1.3.1
sortedcontainers      2.4.0
trio                  0.26.0
trio-websocket        0.11.1
typing_extensions     4.12.2
urllib3               2.2.2
websocket-client      1.8.0
wsproto               1.2.0